Removed automatic dropdown selection of main area
Doesn't make sense to have the drop down selection for the main area be a default.
Can easily be implemented for a give page if desired, but mostly not finding a need for that any more.
Default layout it to have a single main area and a single sidebar now